[persist] Add my-persist-dir
It does not make sense to version control configuration and data files;
they are updated too frequently. However, it is annoying to lose
customizations like bookmarks should we re-clone this
repo.

This commit adds the `my-persist-dir` variable and is used to specify
the directory used to backup files I would like to persist between
sessions -- even if I remove the repo.