[dev] Fix git-commit-mode-hook lambda func
The `setopt` macro was introduced in Emacs29. It similar to `setq` but
slightly different in that it shows a warning when a variable value does
not match `:type` specifications. Effectively, it is a "syntax sugar"
version of `customize-set-variable`.

In this instance, the lambda function was overwriting the previously set
value of `fill-column`. This does not happen when we use `setq`.

https://www.reddit.com/r/emacs/comments/178em7u/why_was_setopt_introduced_in_emacs_29/