[completion] Add text-mode yas-parent snippet
Commit bf8a948 added `git-commit-mode` snippets to formalize commit
messages in the `COMMIT_EDITMSG` buffer. Later discovered that snippets
are unavailable because buffer has `text` major-mode. This was also the
case after installing `yasnippet-snippets` package.

This commit adds .yas-parents -> git-commit-mode to `snippets/text-mode`
directory so that snippets in `git-commit-mode` are available in
`COMMIT_EDITMSG` buffer. The consequence is that the snippets are then
available across all `text-mode` buffers.